chore: update package dependencies and refactor feature flag integration
- Updated package.json to use new versions of OpenFeature SDKs and other dependencies.
- Refactored API routes to utilize the generated OpenFeature client instead of the deprecated server SDK methods.
- Changed imports in checkout and header components to use the new generated hooks for feature flags.
- Created new generated client files for both Node.js and React to streamline feature flag usage.
- Updated database connection logic to align with the new feature flag client methods.
