+++ package.json
@@ -7,9 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "generate-flags:react": "docker run -it -v $(pwd):/local -w /local ghcr.io/open-feature/cli:0.1.10 generate react --flag_manifest_path ./flags.json --output_path ./src/generated/react-hooks.ts",
-    "generate-flags:node": "echo \"flag generation isn't supported for Node.js.\"",
-    "generate-flags": "npm run generate-flags:react && npm run generate-flags:node && echo \"Flags generated.\"",
+    "generate-flags": "make generate-flags",
     "test": "jest"
   },
   "dependencies": {
@@ -17,32 +15,32 @@
     "@openfeature/flagd-core": "^1.0.0",
     "@openfeature/flagd-provider": "^0.13.3",
     "@openfeature/ofrep-web-provider": "^0.3.2",
-    "@openfeature/react-sdk": "^0.4.11",
-    "@openfeature/server-sdk": "^1.17.1",
+    "@openfeature/react-sdk": "^1.0.0",
+    "@openfeature/server-sdk": "^1.18.0",
     "@opentelemetry/api-events": "^0.54.2",
     "@opentelemetry/exporter-logs-otlp-proto": "^0.54.2",
     "@opentelemetry/exporter-metrics-otlp-proto": "^0.54.2",
     "@opentelemetry/exporter-trace-otlp-proto": "^0.54.2",
     "@opentelemetry/sdk-events": "^0.54.2",
-    "@tanstack/react-query": "^5.62.8",
-    "@vercel/otel": "^1.10.1",
-    "lucide-react": "^0.468.0",
-    "next": "14.2.20",
+    "@tanstack/react-query": "^5.74.9",
+    "@vercel/otel": "^1.11.0",
+    "lucide-react": "^0.503.0",
+    "next": "14.2.28",
     "object-hash": "^3.0.0",
-    "react": "^18",
-    "react-dom": "^18",
-    "sharp": "^0.33.5",
-    "uuid": "^11.0.3"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "sharp": "^0.34.1",
+    "uuid": "^11.1.0"
   },
   "devDependencies": {
-    "@tanstack/eslint-plugin-query": "^5.62.1",
+    "@tanstack/eslint-plugin-query": "^5.74.7",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.2.0",
+    "@testing-library/react": "^16.3.0",
     "@types/jest": "^29.5.14",
     "@types/node": "^20",
     "@types/object-hash": "^3.0.6",
-    "@types/react": "^18",
+    "@types/react": "^18.3.20",
     "@types/react-dom": "^18",
     "@types/uuid": "^10.0.0",
     "cross-env": "^7.0.3",
