+++ src/app/api/products/route.ts
@@ -2,11 +2,11 @@
 
 import { NextResponse } from "next/server";
 import { getConnection } from "@/db";
-import { OpenFeature } from "@openfeature/server-sdk";
 import { headerToEvaluationContext } from "@/libs/open-feature/evaluation-context";
+import { getGeneratedClient } from "@/generated/nodejs/openfeature";
 
 export async function GET(req: Request) {
-  const featureFlagClient = OpenFeature.getClient(
+  const featureFlagClient = getGeneratedClient(
     headerToEvaluationContext(req.headers)
   );
   const db = await getConnection(featureFlagClient);
