+++ src/db/index.ts
@@ -1,19 +1,18 @@
-import { Client } from "@openfeature/server-sdk";
 import { Database } from "./types";
 import { PostgresDb } from "./postgres-db";
 import { SqliteDb } from "./sqlite-db";
 import { products } from "./seed";
+import type { GeneratedClient } from "@/generated/nodejs/openfeature";
 
 export type { Database };
 
 const postgres = new PostgresDb({ products });
 const sqlite = new SqliteDb({ products });
 
-export async function getConnection(flagsClient: Client): Promise<Database> {
-  const useDistributed = await flagsClient.getBooleanValue(
-    "use-distributed-db",
-    false
-  );
+export async function getConnection(
+  flagsClient: GeneratedClient
+): Promise<Database> {
+  const useDistributed = await flagsClient.useDistributedDb();
 
   return useDistributed ? postgres : sqlite;
 }
